requirements.txt
# ============================================
# Training Escalade - Dépendances Python
# Application multi-utilisateurs avec MySQL
# ============================================

# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pymysql==1.1.0                    # MySQL driver
cryptography==42.0.0              # Pour PyMySQL SSL

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyjwt==2.8.0

# Email Service (SMTP)
aiosmtplib==3.0.1                 # Async SMTP
email-validator==2.1.0.post1
jinja2==3.1.3                     # Templates emails

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Date & Time
python-dateutil==2.8.2

# Image Processing (photos grandes voies)
pillow==10.2.0

# File Upload
aiofiles==23.2.1

# Utilities
pyyaml==6.0.1

# Testing (optionnel)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development (optionnel)
black==24.1.1                     # Formateur code
flake8==7.0.0                    # Linter